requirements.txt
Django==4.0.3
psycopg2-binary==2.8.6
djangorestframework==3.13.1
markdown==3.3.6
django_extensions==3.1.5
django-filter==21.1